scripts/requirements.txt
opencv-python>=4.7.0
mediapipe>=0.9.0
numpy>=1.21.0
psycopg2-binary>=2.9.0 